Add Selenium Firefox Browser
This commit adds configuration for running Selenium Standalone with
Firefox installed as an add-on container. This allows workflows that
require scripted browser interactions to obtain such a service
by connecting to port 4444 using a selenium client library.

The container builds from the Seleniarm project are used instead
of the official builds as these closely follow upstream but
support aarch64 in addition to amd64.